# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  CreateViewInput:
    properties:
      display_name:
        type: string
      filters:
        items:
          type: integer
        type: array
      slug:
        allOf:
        - $ref: '#/definitions/github_com_generate_selfserve_internal_models.ViewSlug'
        enum:
        - requests_web
        - rooms_web
        - guests_web
    required:
    - filters
    - slug
    type: object
...
  View:
    properties:
      created_at:
        type: string
      display_name:
        type: string
      filters:
        items:
          type: integer
        type: array
      id:
        type: string
      slug:
        $ref: '#/definitions/github_com_generate_selfserve_internal_models.ViewSlug'
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  github_com_generate_selfserve_internal_errs.HTTPError:
    properties:
      code:
        type: integer
      message: {}
    type: object
...
  github_com_generate_selfserve_internal_models.ViewSlug:
    enum:
    - requests_web
    - rooms_web
    - guests_web
    type: string
    x-enum-varnames:
    - ViewSlugRequestsWeb
    - ViewSlugRoomsWeb
    - ViewSlugGuestsWeb
...
host: localhost:8080
info:
  contact:
    email: support@selfserve.com
    name: API Support
  description: SelfServe backend API documentation
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: SelfServe API
  version: "1.0"
paths:
...
  /views:
    get:
      description: Returns all saved filter views for the authenticated user scoped
        to a page slug
      parameters:
      - description: Page slug (e.g. requests_web)
        in: query
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/View'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_generate_selfserve_internal_errs.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_generate_selfserve_internal_errs.HTTPError'
      security:
      - BearerAuth: []
      summary: List views
      tags:
      - views
    post:
      consumes:
      - application/json
      description: Saves the current filter state as a named view for the authenticated
        user
      parameters:
      - description: View payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CreateViewInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/View'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_generate_selfserve_internal_errs.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_generate_selfserve_internal_errs.HTTPError'
      security:
      - BearerAuth: []
      summary: Create view
      tags:
      - views
  /views/{id}:
    delete:
      description: Deletes a saved filter view owned by the authenticated user
      parameters:
      - description: View ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_generate_selfserve_internal_errs.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_generate_selfserve_internal_errs.HTTPError'
      security:
      - BearerAuth: []
      summary: Delete view
      tags:
      - views
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Your Clerk JWT token (prefixed with "Bearer ")
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"